refactor wrgdv3: break roParametersTable into submodules
