+++ src/scripts/modules/gooddata-writer-v3/helpers/inputMapping.js
@@ -1,0 +1,8 @@
+import { Map, fromJS } from 'immutable/dist/immutable';
+const inputMappingPath = ['storage', 'input', 'tables'];
+export function parseInputMapping(configuration) {
+  return configuration.getIn(inputMappingPath.concat(0), Map());
+}
+export function createInputMapping(inputMapping) {
+  return Map().setIn(inputMappingPath, fromJS([inputMapping]));
+}
